array task sum & average
